copytables: a few more warning + a compile fix.
[TESTS]
Compile in 32bit + 64bit locally + Jenkins.

+++ plugins/migration/copytable/copytable.cpp
@@ -251,7 +251,7 @@ RowBuffer::RowBuffer(boost::shared_ptr<std::vector<ColumnInfo> > columns,
       case MYSQL_TYPE_GEOMETRY:
         // source_length is not reliable (and returns bogus value for access)
         // so we just use the max_packet_size value
-        bind.buffer_length = (unsigned long)std::min(max_packet_size, col->source_length + 1);
+        bind.buffer_length = (unsigned long)std::min(max_packet_size, (size_t)col->source_length + 1);
         bind.length = (unsigned long*)malloc(sizeof(unsigned long));
         if (!bind.length)
           throw std::runtime_error("Could not allocate memory for row buffer");
@@ -906,14 +906,14 @@ size_t ODBCCopyDataSource::count_rows(const std::string &schema, const std::stri
   if (!SQL_SUCCEEDED(ret = SQLExecDirect(stmt, (SQLCHAR*)q.c_str(), SQL_NTS)))
     throw ConnectionError("SQLExecDirect("+q+")", ret, SQL_HANDLE_STMT, stmt);
 
-  long long count = 0;
+  size_t count = 0;
   if (SQL_SUCCEEDED(SQLFetch(stmt)))
     SQLGetData(stmt, 1, SQL_C_ULONG, &count, sizeof(count), NULL);
 
   SQLFreeHandle(SQL_HANDLE_STMT, stmt);
 
   if ((spec.type == CopyAll || spec.type == CopyWhere) && spec.max_count > 0 && spec.max_count < count)
-    count = spec.max_count;
+    count = (size_t)spec.max_count;
 
   return count;
 }
@@ -1409,14 +1409,14 @@ size_t MySQLCopyDataSource::count_rows(const std::string &schema, const std::str
   // Retrieves the row count...
   MYSQL_ROW row = mysql_fetch_row(result);
 
-  long long count = 0;
+  size_t count = 0;
   if (row)
     count = atol(row[0]);
 
   mysql_free_result(result);
 
   if ((spec.type == CopyAll || spec.type == CopyWhere) && spec.max_count > 0 && spec.max_count < count)
-      count = spec.max_count;
+      count = (size_t)spec.max_count;
 
   return count;
 }
